+++ database/migrations/2025_10_06_000000_create_user_credits_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
         Schema::create('user_credits', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->uuid('user_id');
-            $table->integer('credits')->default(0);
+            $table->decimal('credits', 10, 2)->default(0.00);
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
